// front/src/providers/api-client.generated/model/notificationDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './userDto';


export interface NotificationDto { 
    id?: string;
    user?: UserDto;
    userId: string;
    creationDate: Date;
    title: string;
    seen: boolean;
    url?: string;
}